[API - URLSession] - Move files into their own folder and make HTTPURLSession conforms to HTTPClient
